+++ config/config.exs
@@ -22,6 +22,15 @@ config :td_qx, TdQxWeb.Endpoint,
   ],
   live_view: [signing_salt: "ZOlKIPdj"]
 
+config :td_qx, Truedat.Auth.Guardian,
+  allowed_algos: ["HS512"],
+  issuer: "tdauth",
+  aud: "truedat",
+  ttl: {1, :hours},
+  secret_key: "SuperSecretTruedat"
+
+config :bodyguard, default_error: :forbidden
+
 # Configures Elixir's Logger
 config :logger, :console,
   format: "$time $metadata[$level] $message\n",