+++ config/runtime.exs
@@ -20,19 +20,7 @@ if System.get_env("PHX_SERVER") do
   config :td_qx, TdQxWeb.Endpoint, server: true
 end
 
-config :td_cluster,
-  scope: :truedat,
-  topologies: [
-    truedat: [
-      strategy: TdCluster.Strategy,
-      config: [
-        node_template: System.get_env("RELEASE_NODE_TEMPLATE", "{{service}}@{{hostname}}"),
-        services: [:dd],
-        groups: [:qx],
-        polling_interval: 10_000
-      ]
-    ]
-  ]
+config :td_cluster, groups: [:qx]
 
 if config_env() == :prod do
   maybe_ipv6 = if System.get_env("ECTO_IPV6") in ~w(true 1), do: [:inet6], else: []
@@ -47,4 +35,6 @@ if config_env() == :prod do
     timeout: System.get_env("DB_TIMEOUT_MILLIS", "15000") |> String.to_integer(),
     ssl: System.get_env("DB_SSL", "") |> String.downcase() == "true",
     socket_options: maybe_ipv6
+
+  config :td_qx, Truedat.Auth.Guardian, secret_key: System.fetch_env!("GUARDIAN_SECRET_KEY")
 end