+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-5809] Authentication and Authorization for admins on DataSets
+
 ## [5.10.0] 2023-07-06
 
 ### Added